Addon Remaster
This commit brings monumental changes to the addon, remaking the block logic, waterlog functionality, and even bringing huge improvements to the way vanilla candles are replaced with working ones. The most major change is support for arbitrarily shaped fill areas around players to replace vanilla candles in. If a group of voxels can be generated, then a greedy grouping algorithm will group them all together according to a given maximum group size. Currently it is set to group to the maximum Dimension.fillBlocks() size, enabling much quicker block replacement functionality for arbitrary volumes! To take advantage of this, the replacement volume is now a specially shaped ellipsoid surrounding all players.